Refactor profile validation
